Initial commit of files.
mklocalroot:

Construct a pristine chroot for building local packages.

local-x86_64-build:

This is a script that mimicks {core,extra}-x86_64-build for building
packages but does so with a preset chroot constructed from mklocalroot.
Local settings are applied and it pushes to a local repo on build success.

needsbuild:

Uses makepkg style tools to check if a package needs building after any
pkgver updating if any.  This doesn't seem to be an option in makepkg
and perhaps a patch for makepkg is in order.

updateall:

Do a git pull on all directories in the current path.  Also, an asp
update is run.

buildall:

Build all packages in the local directory.  Currently assumes a
directory structure like the arch svn build scripts.